HIBERNATE-192: extract Gradle convention pluginsinto buildSrc
Pull the shared java-library, integration-test source set, and publishing/signing
configuration out of the root build into reusable convention plugins, so the new
Spring Boot subprojects can apply them instead of duplicating the setup.
