+++ buildSrc/src/main/kotlin/mongo-hibernate-java.gradle.kts
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2025-present MongoDB, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import org.gradle.api.tasks.testing.logging.TestLogEvent
+
+plugins {
+    id("java-library")
+    id("spotless-java-extension")
+}
+
+java {
+    toolchain { languageVersion = JavaLanguageVersion.of(17) } // Remember to update javadoc links
+    withJavadocJar()
+    withSourcesJar()
+}
+
+tasks.withType<Test>().configureEach {
+    useJUnitPlatform()
+    testLogging { events(TestLogEvent.PASSED, TestLogEvent.SKIPPED, TestLogEvent.FAILED) }
+}
+
+tasks.check { dependsOn(tasks.spotlessApply) }
